# docs/swagger.yaml
basePath: /
definitions:
  auth.authResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  auth.registerRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  category.Category:
    properties:
      id:
        type: integer
      name:
        type: string
      sub_categories:
        items:
          $ref: '#/definitions/category.SubCategory'
        type: array
    type: object
  category.SubCategory:
    properties:
      category_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  expense.Expense:
    properties:
      amount:
        type: number
      category_id:
        type: integer
      category_name:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      sub_category_id:
        type: integer
      sub_category_name:
        type: string
    type: object
  expense.ExpenseUpdate:
    properties:
      amount:
        format: float64
        type: number
      categoryID:
        type: integer
      description:
        type: string
      subCategoryID:
        type: integer
    type: object
  report.CategoryTotal:
    properties:
      category:
        type: string
      expenses:
        items:
          $ref: '#/definitions/expense.Expense'
        type: array
      total:
        type: number
    type: object
  report.DayTotal:
    properties:
      day:
        type: string
      expenses:
        items:
          $ref: '#/definitions/expense.Expense'
        type: array
      total:
        type: number
    type: object
  report.MonthTotal:
    properties:
      expenses:
        items:
          $ref: '#/definitions/expense.Expense'
        type: array
      month:
        type: string
      total:
        type: number
    type: object
  rest.categoryRequest:
    properties:
      name:
        type: string
    type: object
  rest.dashboardResponse:
    properties:
      by_category:
        items:
          $ref: '#/definitions/report.CategoryTotal'
        type: array
      by_day:
        items:
          $ref: '#/definitions/report.DayTotal'
        type: array
      by_month:
        items:
          $ref: '#/definitions/report.MonthTotal'
        type: array
      expenses:
        items:
          $ref: '#/definitions/expense.Expense'
        type: array
    type: object
  rest.paginatedExpensesResponse:
    properties:
      expenses:
        items:
          $ref: '#/definitions/expense.Expense'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
    type: object
info:
  contact: {}
  title: Expense Tracker API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.authResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - auth
  /categories:
    get:
      description: Returns the authenticated user's categories with nested sub-categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/category.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category name
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/rest.categoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Existing expenses referencing this category will have their category
        set to null
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: New name
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/rest.categoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Rename a category
      tags:
      - categories
  /categories/{id}/sub-categories:
    post:
      consumes:
      - application/json
      parameters:
      - description: Parent category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Sub-category name
        in: body
        name: subcategory
        required: true
        schema:
          $ref: '#/definitions/rest.categoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/category.SubCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a sub-category under a category
      tags:
      - categories
  /dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.dashboardResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Summary
  /expenses:
    get:
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Search by description (case-insensitive substring match)
        in: query
        name: search
        type: string
      - description: Filter from this date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Filter up to this date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.paginatedExpensesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Expenses
    post:
      consumes:
      - application/json
      parameters:
      - description: Expense to create
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/expense.Expense'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/expense.Expense'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Expenses
  /expenses/{id}:
    delete:
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Expenses
    get:
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/expense.Expense'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Expenses
    put:
      consumes:
      - application/json
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: updates
        required: true
        schema:
          $ref: '#/definitions/expense.ExpenseUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/expense.Expense'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Expenses
  /expenses/export:
    get:
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Search by description (case-insensitive substring match)
        in: query
        name: search
        type: string
      - description: Filter from this date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Filter up to this date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Expenses
  /health:
    get:
      description: Confirms the API is responsive and that PostgreSQL is reachable.
      produces:
      - application/json
      responses:
        "200":
          description: API and database are healthy
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: API is reachable but database is unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check API health
      tags:
      - Health
  /sub-categories/{id}:
    delete:
      parameters:
      - description: Sub-category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a sub-category
      tags:
      - categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Sub-category ID
        in: path
        name: id
        required: true
        type: integer
      - description: New name
        in: body
        name: subcategory
        required: true
        schema:
          $ref: '#/definitions/rest.categoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.SubCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Rename a sub-category
      tags:
      - categories
  /summary/category:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/report.CategoryTotal'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Summary
  /summary/day:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/report.DayTotal'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Summary
  /summary/month:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/report.MonthTotal'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Summary
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"